Fixed issue with readability in volley response listener. We now take in the string response, and simple convert it into a json object, we can then parse it using a get string for the information we need.
A user is now able to login and have their user information initalized such as userid, password, username, bio etc. This information is stored in a user information singleton class that is only allowed to have a single instance running at once.

This instance will be able to be used across view models and the rest of the activity to quickly grab user information and use it as needed.