Play Fragment - Now hooked up to gameWFriends Activity, will allow users to launch the game.
GameWFriendsActivity - Currently has one button with no functionality, button will be used to pull a screen overlay of a websocket chat.

New Package created called Game, and within it is the gamewfriends activity, as well as a ui for chat,

Chat Fragment - will be used to bring in a chat screen to be used via websocket.

ChatTools - Initalized all the methods and basic structure for implementing a websocket chat using okhttp.

Added dependency of okhttp, as well as updated android manifest, removing old stuff that we no longer have as well as adding the new GameWFriendsActivity.

Other basic javadocs, and bug fixes.